Изменил логику отрисовки header'a
...по аналогии с футером
