+++ public/index.html
@@ -12,18 +12,9 @@
 </head>
 <body>
 
-<header class="header">
-  <h1 class="header__logo logo">Cinemaddict</h1>
-
-  <!-- Звание пользователя -->
-</header>
-
 <main class="main">
 </main>
 
-
-<!-- <footer class="footer">
-</footer> -->
 <script src="./bundle.js"></script>
 </body>
 </html>
