+++ src/main.js
@@ -4,12 +4,11 @@ import HeaderPresenter from './presenter/header-presenter.js';
 import MainPresenter from './presenter/main-presenter.js';
 
 const siteBody = document.body;
-const siteHeaderElement = document.querySelector('.header');
 const siteMainElement = document.querySelector('.main');
 
 const cardsModel = new CardsModel();
 
-const headerPresenter = new HeaderPresenter(siteHeaderElement);
+const headerPresenter = new HeaderPresenter(siteBody);
 headerPresenter.init();
 
 const mainPresenter = new MainPresenter(siteMainElement, cardsModel);
